Feature: VB-3763 - Prisoner alert change events are processed
Introduce the service logic to handle pulling back effected visits
if an alert code is found to be within the supported alert codes enum set.

Add TODO comment for reworking the saving of the notification to include a
new field description which will hold a detailed message for the staff member
to review the reason a visit needs managing.