Feature: VB-3763 - Prisoner alert change events are processed
Update the event_audit table to include a nullable column description.

In the description column will go a detailed message of the event and will later
be displayed to the staff during proactive booking management. It will help them
understand why the booking has been flagged.

Update alert endpoint integration tests to assert the description column is working
as expected