Feature: VB-3763 - Prisoner alert change events are processed
After discussion the enum for alert codes has been stripped to only
include the codes. The other information isn't something the API
should be concerned with.

Also add to the DTO a new description field which will be passed in
via the request. This description will be stored in the visit_notification_event
table to give more information on why a visit was flagged.