Feature: VB-2761 - Prisoner transferred events are processed (#723)
* Feature: VB-2761 - Handle prisoner received notification reason TRANSFERRED

When a prisoner is received, handle the notification for reason TRANSFERRED.

Update the Dto to include the reason of receiving prisoner.
Update the service layer to implement the handlePrisonerReceivedNotification for reason TRANSFERRED.
Update NotificationEventType to include new event PRISONER_RECEIVED_EVENT.
Update and add enums for changes.

* Feature: VB-2761 - Prisoner transferred events are processed

Changes due to testing. The prison code provided by the SQS message
is the new prison that the prisoner has moved to. For the flagged
visits to be found we need the previous prison (where the visits are
booked for).

Change to use lastPrisonId instead of current prison.

* Feature: VB-2761 - Prisoner transferred events are processed

Rename the lastPrisonId to lastPrisonCode to keep with naming convention
within our service, as per pr comments.

* Feature: VB-2761 - Prisoner transferred events are processed

Account for a potentially null lastPrisonCode, if they don't have a
lastPrisonCode then we can't find the visits. So do nothing.