Add MIGRATED_VISIT as event audit to search in query to get last user to update booking by reference
A bug in production is showing where result set of this query is coming back null.
This is due to the visit being a migrated visit and not having a BOOKED_VISIT event in the event audit table.
It also hasn't been updated so it doesn't have an UPDATED_VISIT either.

Adding MIGRATED_VISIT to the query to make sure it never comes back null.