When frequency of session template is more than 4 weeks it was causing a bug of not displaying the session template
This was due to how we aligned the date with the closest day of the week matching the template
and the booking window start date.

Rework the logic to first align the date and then loop through using the frequency to get first future date.

Currently this is a crude fix, we're looking into optimising as templates can be created many years ago and this
wouldn't be suitable for those due to the while loop constantly checking.