VB-4514 - Include additional information for visitor specific events
When an event type is for a visitor, capture the additional information
from that event in the visit_notification_event table.

VisitorId - The ID of the visitor who has had an event raised about them
VisitorRestrictionType - For the visitor, what restriction has been added

Update the NotificationGroupDto to include new fields, ready to be returned to
the orchestration and frontend.